+++ docs-util/packages/typedoc-config/entities.js
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const path = require("path")
-const globalTypedocOptions = require("./base")
+const globalTypedocOptions = require("./_base")
 
 const pathPrefix = path.join(__dirname, "..", "..", "..")
 
@@ -11,11 +11,31 @@ module.exports = {
   tsconfig: path.join(pathPrefix, "packages/medusa/tsconfig.json"),
   name: "Entities Reference",
   indexTitle: "Entities Reference",
-  entryDocument: "index.md",
+  entryDocument: "_index.mdx",
   hideInPageTOC: true,
   hideBreadcrumbs: true,
   plugin: [...globalTypedocOptions.plugin, "typedoc-plugin-frontmatter"],
   frontmatterData: {
     displayed_sidebar: "entitiesSidebar",
   },
+  formatting: {
+    "*": {
+      showCommentsAsHeader: true,
+      sections: {
+        member_sources_definedIn: false,
+        reflection_hierarchy: false,
+      },
+      parameterStyle: "component",
+      parameterComponent: "ParameterTypes",
+      mdxImports: [
+        `import ParameterTypes from "@site/src/components/ParameterTypes"`,
+      ],
+      reflectionGroups: {
+        Constructors: false,
+        Methods: false,
+      },
+    },
+  },
+  objectLiteralTypeDeclarationStyle: "component",
+  mdxOutput: true,
 }