+++ packages/medusa/src/loaders/helpers/routing/__fixtures__/routers-middleware/customers/route.ts
@@ -1,0 +1,9 @@
+import { Request, Response } from "express"
+
+export function GET(req: Request, res: Response) {
+  res.send("list customers")
+}
+
+export function POST(req: Request, res: Response) {
+  res.send("create customer")
+}