+++ packages/types/src/pricing/common/rule-type.ts
@@ -3,12 +3,12 @@ import { BaseFilterable } from "../../dal"
 /**
  * @interface
  * 
- * An object that holds the details of a rule type.
+ * A rule type's data.
  * 
- * @prop id - A string indicating the ID of the rule type.
- * @prop name - A string indicating the display name of the rule type.
- * @prop rule_attribute - A string indicating a unique name used to later identify the rule_attribute. For example, it can be used in the `context` parameter of the `calculatePrices` method to specify a rule for calculating the price.
- * @prop default_priority - A number indicating the priority of the rule type. This is useful when calculating the price of a price set, and multiple rules satisfy the provided context. The higher the value, the higher the priority of the rule type.
+ * @prop id - The ID of the rule type.
+ * @prop name - The display name of the rule type.
+ * @prop rule_attribute - The unique name used to later identify the rule_attribute. For example, it can be used in the `context` parameter of the `calculatePrices` method to specify a rule for calculating the price.
+ * @prop default_priority - The priority of the rule type. This is useful when calculating the price of a price set, and multiple rules satisfy the provided context. The higher the value, the higher the priority of the rule type.
  */
 export interface RuleTypeDTO {
   id: string
@@ -20,12 +20,12 @@ export interface RuleTypeDTO {
 /**
  * @interface
  * 
- * An object used when creating a rule type to specify its data.
+ * The rule type to create.
  * 
- * @prop id - A string indicating the ID of the rule type.
- * @prop name - A string indicating the display name of the rule type.
- * @prop rule_attribute - A string indicating a unique name used to later identify the rule_attribute. For example, it can be used in the `context` parameter of the `calculatePrices` method to specify a rule for calculating the price.
- * @prop default_priority - A number indicating the priority of the rule type. This is useful when calculating the price of a price set, and multiple rules satisfy the provided context. The higher the value, the higher the priority of the rule type.
+ * @prop id - The ID of the rule type.
+ * @prop name - The display name of the rule type.
+ * @prop rule_attribute - The unique name used to later identify the rule_attribute. For example, it can be used in the `context` parameter of the `calculatePrices` method to specify a rule for calculating the price.
+ * @prop default_priority - The priority of the rule type. This is useful when calculating the price of a price set, and multiple rules satisfy the provided context. The higher the value, the higher the priority of the rule type.
  */
 export interface CreateRuleTypeDTO {
   id?: string
@@ -37,12 +37,12 @@ export interface CreateRuleTypeDTO {
 /**
  * @interface
  * 
- * An object used when updating a rule type to specify the data to update.
+ * The data to update in a rule type. The `id` is used to identify which price set to update.
  * 
- * @prop id - A string indicating the ID of the rule type to update.
- * @prop name - A string indicating the display name of the rule type.
- * @prop rule_attribute - A string indicating a unique name used to later identify the rule_attribute. For example, it can be used in the `context` parameter of the `calculatePrices` method to specify a rule for calculating the price.
- * @prop default_priority - A number indicating the priority of the rule type. This is useful when calculating the price of a price set, and multiple rules satisfy the provided context. The higher the value, the higher the priority of the rule type.
+ * @prop id - The ID of the rule type to update.
+ * @prop name - The display name of the rule type.
+ * @prop rule_attribute - The unique name used to later identify the rule_attribute. For example, it can be used in the `context` parameter of the `calculatePrices` method to specify a rule for calculating the price.
+ * @prop default_priority - The priority of the rule type. This is useful when calculating the price of a price set, and multiple rules satisfy the provided context. The higher the value, the higher the priority of the rule type.
  */
 export interface UpdateRuleTypeDTO {
   id: string
@@ -54,11 +54,11 @@ export interface UpdateRuleTypeDTO {
 /**
  * @interface
  * 
- * An object used to filter retrieved rule types.
+ * Filters to apply on rule types.
  * 
- * @prop id - an array of strings, each being an ID to filter rule types.
- * @prop name - an array of strings, each being a name to filter rule types.
- * @prop rule_attribute - an array of strings, each being a rule attribute to filter rule types.
+ * @prop id - The IDs to filter rule types by.
+ * @prop name - The names to filter rule types by.
+ * @prop rule_attribute - The rule attributes to filter rule types by.
  */
 export interface FilterableRuleTypeProps
   extends BaseFilterable<FilterableRuleTypeProps> {