+++ packages/medusa-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 9.0.9
+
+### Patch Changes
+
+- [`045d1b6a0`](https://github.com/medusajs/medusa/commit/045d1b6a0c2b0d03e5fa1db886d1f81c843059ce) Thanks [@olivermrbl](https://github.com/olivermrbl)! - chore: Bump @medusajs/medusa dep
+
+- Updated dependencies [[`045d1b6a0`](https://github.com/medusajs/medusa/commit/045d1b6a0c2b0d03e5fa1db886d1f81c843059ce)]:
+  - @medusajs/medusa-js@6.1.3
+
+## 9.0.8
+
+### Patch Changes
+
+- [#5400](https://github.com/medusajs/medusa/pull/5400) [`5a5c96e21`](https://github.com/medusajs/medusa/commit/5a5c96e2118e50a558d9d6dc340e505454d4c593) Thanks [@dwene](https://github.com/dwene)! - add types to exports to help projects using moduleResolution bundler
+
+- Updated dependencies [[`5a5c96e21`](https://github.com/medusajs/medusa/commit/5a5c96e2118e50a558d9d6dc340e505454d4c593), [`98e275551`](https://github.com/medusajs/medusa/commit/98e275551415583602763cf457c3f95400209d0a)]:
+  - @medusajs/medusa-js@6.1.2
+
 ## 9.0.7
 
 ### Patch Changes